feat: enhance validation
