+++ product-service/handler.ts
@@ -1,7 +1,1 @@
-import { getAllProducts } from "./src/get-all-products";
-import { getProductById } from "./src/get-product-by-id";
-import { getStaticImage } from "./src/get-static-image";
-
-import "source-map-support/register";
-
-export { getProductById, getAllProducts, getStaticImage };
+export * from './src/handlers';
