+++ product-service/src/handler/get-product-by-id.ts
@@ -1,20 +1,20 @@
 import { APIGatewayProxyHandler } from "aws-lambda";
 import "source-map-support/register";
 
-import { productDb } from "./product-db";
-import { httpResponse } from "./libs";
+import { productService } from "../services/product-service";
+import { httpResponse } from "../libs/http";
 
 type Params = {
   productId?: string;
 };
 
-const isInvalidParams = (p: Params) => !p["productId"];
+const isInvalidParams = (p: Params) => !p.hasOwnProperty("productId")
 
 export const getProductById: APIGatewayProxyHandler = async (
   event,
   _context
 ) => {
-  const params = event.pathParameters || {};
+  const params = Object(event.pathParameters);
 
   console.info(event);
 
@@ -25,7 +25,7 @@ export const getProductById: APIGatewayProxyHandler = async (
   const productId = params["productId"];
 
   try {
-    const product = await productDb.selectProductById(productId);
+    const product = await productService.getProductById(productId);
 
     if (!product) {
       return httpResponse.failure({
@@ -33,7 +33,7 @@ export const getProductById: APIGatewayProxyHandler = async (
       });
     }
 
-    return httpResponse.success({ data: product || null });
+    return httpResponse.success({ data: product });
   } catch (e) {
     console.error(e.message);
