+++ product-service/src/handlers.ts
@@ -1,0 +1,7 @@
+import { getAllProducts } from "./handler/get-all-products";
+import { getProductById } from "./handler/get-product-by-id";
+import { getStaticAsset } from "./handler/get-static-asset";
+
+import "source-map-support/register";
+
+export { getProductById, getAllProducts, getStaticAsset };
